Playerの動作を追加

+++ BattleCatsLike/Assets/Scripts/Character/ICharacterBehavior.cs
@@ -1,7 +1,7 @@
 
 public interface ICharacterBehavior
 {
-    int PositionX { get; set; }
+    float PositionX { get; set; }
 
     //ƒ_ƒ[ƒW‚ğ—^‚¦‚é
     void Damage(int damage);
